making sure everything is long id

+++ src/main/java/com/example/demo/service/EmployeeServicelmpl.java
@@ -1,5 +1,6 @@
 package com.example.demo.service;
 
 public interface EmployeeServicelmpl {
-    void deleteById(int theId);
+    void deleteById(long theId);
+
 }
